pyproject.toml
[project]
name = "argus"
version = "2.0.0"
description = "LLM-based agent for general purpose tasks"
authors = [
  { name ="Ali Çağlar Koçer" },
  { name = "Ali Çağlar Koçer", email = "longrongchen@163.com" },
  { name = "Ali Çağlar Koçer", email = "yiboxue@163.com" },
  { name = "Ali Çağlar Koçer"},
]
license = {text = "MIT"}
keywords = ["agent", "cli", "llm"]
classifiers = [
    "Programming Language:: Python:: 3",
    "License:: OSI Approved:: MIT License",
    "Operating System:: OS Independent",
]
readme = "README.md"
requires-python = ">=3.10, <3.13"
dependencies = [
    "openai>=2.6.0",
    "anthropic>=0.54.0",
    "click>=8.0.0",
    "google-genai>=1.24.0",
    "jsonpath-ng>=1.7.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "typing-extensions>=4.0.0",
    "ollama>=0.5.1",
    "socksio>=1.0.0",
    "tree-sitter-languages==1.10.2",
    "tree-sitter==0.21.3",
    "ruff>=0.12.4",
    "keyboard>=0.13.5",
    "textual>=5.0.1",
    "prompt-toolkit>=3.0.51",
    "aiohttp>=3.12.13",
    "loguru>=0.7.3",
    "mcp>=1.13.0",
    "pytest>=8.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
evaluation = [
    "sb-cli>=0.1.4",
    "docker>=7.0.0",
    "tqdm>=4.66.0",
    "datasets>=2.14.0",
]

[project.scripts]
argus = "argus.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["argus"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = [
    "B",
    "SIM",
    "C4",
    "E4", "E9", "E7", "F",
    "I"
]

